requirements.txt
boto3
pillow
imagehash
numpy
scikit-image
opencv-python
pytesseract
requests
scikit-learn
